% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_allom.R
\name{calculate_allom}
\alias{calculate_allom}
\title{Calculate tree biomass or volume over bark based on data from tree inventories and allometries from AllometrApp}
\usage{
calculate_allom(
  df,
  variable,
  equation = NULL,
  dbh = "dbh",
  height = NULL,
  sps,
  frac = FALSE
)
}
\arguments{
\item{df}{The dataframe that contains the data}

\item{variable}{The dependent variable that we want to calculate. "BAT" stands for aerial biomass, and "VOB" for volume over bark}

\item{equation}{The allometric equation we want to use. Options include "INIA", "IEFC_1", "IEFC_2", "RUIZ", and "VOB"}

\item{dbh}{The name of the variable that identifies diameter in the dataset}

\item{height}{The name of the variable that identifies tree height in the dataset. If \code{height} is missing, the funcion will calculate it from allometries and warn about it.}

\item{sps}{A variable containing the names of the species}

\item{frac}{Whether we want to obtain the biomass values by tree fraction (leaves, small and large branches, stem, bark). By default, \code{frac = FALSE}}
}
\value{
\itemize{
\item If \code{frac = FALSE}, it will return a numeric vector containing the BAT or VOB values for each observation in \code{df}.
\item If \code{frac = TRUE}, it will return a data frame identical to \code{df} but with some extra columns containing the biomass weights per tree fraction
}
}
\description{
Calculate tree biomass or volume over bark based on data from tree inventories and allometries from AllometrApp
}
\examples{
data(inventory)
calculate_allom(inventory, "BAT", equation = "INIA", dbh = "Dn", height = "Ht", sps = "Sps_name")  # including height
calculate_allom(inventory, "BAT", equation = "IEFC_2", dbh = "Dn", sps = "Sps_name")  # without height
calculate_allom(inventory, "BAT", equation = "RUIZ", dbh = "Dn", sps = "Sps_name", frac = T)  # biomass per fractions

}
