% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ba.R
\name{get_BA}
\alias{get_BA}
\title{Calculates dasometric summaries for IFN plots}
\usage{
get_BA(
  IFN,
  df,
  province = NULL,
  com_aut = NULL,
  species = NULL,
  per_CD = FALSE,
  per_species = FALSE
)
}
\arguments{
\item{df}{data frame}

\item{province}{vector of characters containing the name of the provinces to be included in the analyses}

\item{com_aut}{vector of characters containing the name of the autonomous regions (Comunidad Autónoma) to be included in the analyses}

\item{species}{vector of characters containing the name of the tree species to be included in the analyses}

\item{per_CD}{optional. When `per_CD = TRUE`, the desired variable is summarized per plot and diameter class}

\item{per_species}{optional. When `per_species = TRUE`, the desired variable is summarized per plot and species}
}
\value{
A data frame containing the level of aggregation specified (per plot, per plot & species, per plot and DC, per plot, species and DC), and the amount of the outùt variable specified
}
\description{
Calculates dasometric summaries for IFN plots
}
